# spec/openapi.yaml
openapi: 3.0.3
info:
  title: MealForward B2B AX Platform API
  description: >
    LLM Guardrail AX 기반 세무 리스크 없는 B2B 식대 이월 및 복지 정산 플랫폼 API 명세서.
    비과세 식대 포인트의 이월/사용/정산 흐름과, 현금성 자산 결제를 자동 차단하는
    AI 세무 가드레일(Guardrail AX) 엔드포인트를 포함한다.
  version: 1.0.0
servers:
  - url: http://localhost:3000/api
    description: Local Dev Server
...
paths:
  /auth/signup/user:
    post:
      tags: [Auth]
      summary: 임직원 회원가입
      operationId: signupUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, password]
              properties:
                name:
                  type: string
                  example: "김정인"
                email:
                  type: string
                  example: "jeongin.kim@example.com"
                password:
                  type: string
                  example: "demo1234"
                department:
                  type: string
                  example: "플랫폼개발팀"
                employeeNo:
                  type: string
                  example: "E2024-104"
      responses:
        '200':
          description: 가입 완료
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthProfile'
        '409':
          description: 이미 가입된 이메일
  /auth/signup/admin:
    post:
      tags: [Auth]
      summary: 관리자 회원가입 (관리자 인증 키 필요)
      description: >
        adminKey가 서버의 ADMIN_SIGNUP_KEY와 일치해야만 가입이 승인된다.
        관리자도 임직원이므로, 아직 같은 이메일의 임직원 계정이 없다면 동일한 이메일/비밀번호로
        포인트 상점 이용이 가능한 임직원 계정을 함께 생성한다(반대 방향은 하지 않음).
      operationId: signupAdmin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, password, adminKey]
              properties:
                name:
                  type: string
                  example: "정관리"
                email:
                  type: string
                  example: "admin@example.com"
                password:
                  type: string
                  example: "admin1234"
                adminKey:
                  type: string
                  description: 회사에서 발급한 관리자 인증 키 (ADMIN_SIGNUP_KEY)
                  example: "MEALFORWARD-ADMIN-2026"
      responses:
        '200':
          description: 가입 완료
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthProfile'
        '403':
          description: 관리자 인증 키 불일치
        '409':
          description: 이미 가입된 이메일
  /auth/login:
    post:
      tags: [Auth]
      summary: 임직원/관리자 로그인
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, role]
              properties:
                email:
                  type: string
                  example: "jeongin.kim@example.com"
                password:
                  type: string
                  example: "demo1234"
                role:
                  type: string
                  enum: [user, admin]
                  example: "user"
      responses:
        '200':
          description: 로그인 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthProfile'
        '401':
          description: 이메일 또는 비밀번호 불일치
...
  /admin/settlements/summary:
    get:
      tags: [Admin]
      security:
        - AdminSession: []
      summary: "[Admin] 전사 이월 정산 및 LLM Guardrail 차단 현황 조회"
      operationId: getAdminSummary
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  fiscalYear:
                    type: string
                    example: "2026"
                  totalRolledOverAmount:
                    type: number
                    example: 45200000.00
                  blockedCountByAI:
                    type: integer
                    example: 142
                  savedTaxRiskCost:
                    type: number
                    example: 9040000.00
  /admin/settlements/list:
    get:
      tags: [Admin]
      security:
        - AdminSession: []
      summary: "[Admin] 부서별/직원별 이월 내역 조회"
      operationId: getSettlementList
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    userId:
                      type: integer
                    name:
                      type: string
                    department:
                      type: string
                    rolledAmount:
                      type: number
                    status:
                      type: string
                      example: "ACTIVE"
  /admin/guardrail-policy:
    get:
      tags: [Admin]
      security:
        - AdminSession: []
      summary: "[Admin] 가드레일 정책(현금성 자산 차단) 조회"
      operationId: getGuardrailPolicy
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GuardrailPolicy'
    put:
      tags: [Admin]
      security:
        - AdminSession: []
      summary: "[Admin] 가드레일 정책 갱신 (기본 ON 고정, 세부 민감도만 조정 가능)"
      operationId: updateGuardrailPolicy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GuardrailPolicy'
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GuardrailPolicy'
  /admin/settlements/finalize:
    post:
      tags: [Admin]
      security:
        - AdminSession: []
      summary: "[Admin] 9월 말 연간 정산 최종 승인 및 ERP/회계 연동"
      operationId: finalizeSettlement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                fiscalYear:
                  type: string
                  example: "2026"
      responses:
        '200':
          description: 정산 승인 및 ERP 전송 완료
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  fiscalYear:
                    type: string
                    example: "2026"
                  erpSyncStatus:
                    type: string
                    example: "SYNCED"
  /admin/products:
    get:
      tags: [Admin]
      security:
        - AdminSession: []
      summary: "[Admin] 복지 상품 목록 조회"
      operationId: getAdminProducts
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
    post:
      tags: [Admin]
      security:
        - AdminSession: []
      summary: "[Admin] 복지 상품 등록"
      operationId: createProduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, category, price]
              properties:
                name:
                  type: string
                  example: "온라인 강의 30% 할인 쿠폰"
                category:
                  type: string
                  enum: [BOOK, LEARNING, FITNESS, CAFE, GIFT_CARD]
                price:
                  type: number
                  example: 15000.00
                isCashLike:
                  type: boolean
                  example: false
      responses:
        '200':
          description: 등록 완료
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          description: 필수 필드 누락
  /admin/products/{id}:
    put:
      tags: [Admin]
      security:
        - AdminSession: []
      summary: "[Admin] 복지 상품 수정"
      operationId: updateProduct
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                category:
                  type: string
                  enum: [BOOK, LEARNING, FITNESS, CAFE, GIFT_CARD]
                price:
                  type: number
                isCashLike:
                  type: boolean
      responses:
        '200':
          description: 수정 완료
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: 상품 없음
    delete:
      tags: [Admin]
      security:
        - AdminSession: []
      summary: "[Admin] 복지 상품 삭제"
      operationId: deleteProduct
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 삭제 완료
        '404':
          description: 상품 없음
...
components:
  securitySchemes:
    AdminSession:
      type: apiKey
      in: header
      name: x-admin-id
      description: 로그인한 관리자의 id. /auth/login(role=admin) 응답의 id 값을 그대로 전달한다.
  schemas:
    AuthProfile:
      type: object
      properties:
        role:
          type: string
          example: "user"
        id:
          type: integer
          example: 1004
        name:
          type: string
          example: "김정인"
        department:
          type: string
          example: "플랫폼개발팀"
        employeeNo:
          type: string
          example: "E2024-104"
...
    Product:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "온라인 강의 30% 할인 쿠폰"
        category:
          type: string
          example: "LEARNING"
        price:
          type: number
          example: 15000.00
        isCashLike:
          type: boolean
          description: 상품권/현금성 자산 여부 (true면 식대 이월 포인트로는 결제 불가, 일반 복지 포인트로는 결제 가능)
          example: false
    GuardrailPolicy:
      type: object
      properties:
        blockGiftCards:
          type: boolean
          example: true
        blockCashEquivalents:
          type: boolean
          example: true
        riskThreshold:
          type: number
          description: 이 값 이상의 risk score는 자동 차단
          example: 0.7
...